+++ src/win/processx-win.h
@@ -19,12 +19,8 @@ typedef struct processx_handle_s {
 int processx__utf8_to_utf16_alloc(const char* s, WCHAR** ws_ptr);
 
 int processx__stdio_create(processx_handle_t *handle,
-			   HANDLE* extra_connections, int num_connections,
-			   const char *std_in, const char *std_out,
-			   const char *std_err,
-			   BYTE** buffer_ptr, SEXP privatex,
-			   const char *encoding,
-                           const char *cname);
+			   SEXP connections, BYTE** buffer_ptr, SEXP privatex,
+			   const char *encoding, const char *cname, int* inherit_std);
 WORD processx__stdio_size(BYTE* buffer);
 HANDLE processx__stdio_handle(BYTE* buffer, int fd);
 void processx__stdio_destroy(BYTE* buffer);