+++ src/win/utils.c
@@ -15,3 +15,8 @@ SEXP processx__echo_off() {
   R_THROW_ERROR("Only implemented on Unix");
   return R_NilValue;
 }
+
+SEXP processx_make_fifo(SEXP name) {
+  /* TODO */
+  return R_NilValue;
+}