+++ NEWS.md
@@ -5,6 +5,10 @@
   process if you include `"current"` in the value of `env`, in `run()`
   and for `process$new()`: `env = c("current", NEW = "newvalue")` (#232).
 
+* Sub-processes can now inherit the standard input, output and error from
+  the main R process, by setting the corresponding argument to an empty
+  string. E.g. `run("ls", stdout = "")` (#72).
+
 * `run()` is now much faster with large standard output or standard
   error (#286).